+++ src/app/today-news/detail/[id]/page.tsx
@@ -1,5 +1,5 @@
 // 클라이언트 컴포넌트에 파라미터(id)를 전달하는 서버 컴포넌트
-import TodayNewsDetail from "@/features/today-news/components/TodayNewsDetail";
+import TodayNewsDetail from "@/features/detail/components/TodayNewsDetail";
 import { use } from "react";
 
 type Props = {
@@ -7,7 +7,7 @@ type Props = {
 };
 
 export default function TodayNewsDetailPage(props: Props) {
-  const params = use(props.params); // Promise 해제 -> { id: string }
+  const params = use(props.params); 
   const id = Number(params.id);
 
   return <TodayNewsDetail id={id} />;
